+++ engine/crates/engine/Cargo.toml
@@ -36,7 +36,6 @@ futures-util = { workspace = true, features = ["io", "sink"] }
 futures-channel = { version = "0.3.30", default-features = false }
 gateway-v2-auth-config.workspace = true
 http.workspace = true
-im = "15.1.0"
 indexmap.workspace = true
 indoc = "2"
 Inflector = { version = "0.11" }
@@ -60,6 +59,7 @@ web-time.workspace = true
 dataloader.workspace = true
 engine-derive = { path = "derive", version = "3" }
 engine-parser = { path = "parser", version = "3" }
+engine-response.workspace = true
 engine-scalars.path = "scalars"
 engine-validation.workspace = true
 engine-value = { path = "value", version = "3" }
@@ -67,6 +67,7 @@ grafbase-tracing.workspace = true
 graph-entities.workspace = true
 graphql-cursor.workspace = true
 meta-type-name.path = "meta-type-name"
+query-path.workspace = true
 registry-v1.workspace = true
 registry-v2.workspace = true
 runtime.workspace = true
