+++ engine/crates/engine/response/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "engine-response"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+http.workspace = true
+
+thiserror.workspace = true
+common-types.workspace = true
+engine-parser.workspace = true
+engine-value.workspace = true
+grafbase-tracing.workspace = true
+graph-entities.workspace = true
+query-path.workspace = true
+runtime.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+strum.workspace = true
+strum_macros.workspace = true
