+++ engine/crates/engine/src/error.rs
@@ -1,16 +1,18 @@
 use std::{
     any::Any,
-    collections::BTreeMap,
     convert::Infallible,
     fmt::{self, Debug, Display, Formatter},
     marker::PhantomData,
     sync::Arc,
 };
 
-use serde::{Deserialize, Serialize};
+use engine_response::error::ErrorExtensionValues;
+use engine_value::ConstValue;
 use thiserror::Error;
 
-use crate::{parser, LegacyInputType, Pos, QueryPathSegment, Value};
+pub use engine_response::error::ServerError;
+
+use crate::{LegacyInputType, Pos};
 
 #[derive(Debug, serde::Serialize, serde::Deserialize, strum_macros::Display)]
 #[serde(rename_all = "SCREAMING_SNAKE_CASE")]
@@ -18,148 +20,6 @@ use crate::{parser, LegacyInputType, Pos, QueryPathSegment, Value};
 pub enum ErrorCode {
     PersistedQueryNotFound,
     InternalServerError,
-}
-
-/// Extensions to the error.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Default)]
-#[serde(transparent)]
-pub struct ErrorExtensionValues(pub BTreeMap<String, Value>);
-
-impl ErrorExtensionValues {
-    /// Set an extension value.
-    pub fn set(&mut self, name: impl AsRef<str>, value: impl Into<Value>) {
-        self.0.insert(name.as_ref().to_string(), value.into());
-    }
-
-    /// Unset an extension value.
-    pub fn unset(&mut self, name: impl AsRef<str>) {
-        self.0.remove(name.as_ref());
-    }
-}
-
-/// An error in a GraphQL server.
-#[derive(Clone, Serialize, Deserialize)]
-pub struct ServerError {
-    /// An explanatory message of the error.
-    pub message: String,
-    /// The source of the error.
-    #[serde(skip)]
-    pub source: Option<Arc<dyn Any + Send + Sync>>,
-    /// Where the error occurred.
-    #[serde(skip_serializing_if = "Vec::is_empty", default)]
-    pub locations: Vec<Pos>,
-    /// If the error occurred in a resolver, the path to the error.
-    #[serde(skip_serializing_if = "Vec::is_empty", default)]
-    pub path: Vec<QueryPathSegment>,
-    /// Extensions to the error.
-    #[serde(skip_serializing_if = "error_extensions_is_empty", default)]
-    pub extensions: Option<ErrorExtensionValues>,
-}
-
-fn error_extensions_is_empty(values: &Option<ErrorExtensionValues>) -> bool {
-    values.as_ref().map_or(true, |values| values.0.is_empty())
-}
-
-impl Debug for ServerError {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        f.debug_struct("ServerError")
-            .field("message", &self.message)
-            .field("locations", &self.locations)
-            .field("path", &self.path)
-            .field("extensions", &self.extensions)
-            .finish_non_exhaustive()
-    }
-}
-
-impl PartialEq for ServerError {
-    fn eq(&self, other: &Self) -> bool {
-        self.message.eq(&other.message)
-            && self.locations.eq(&other.locations)
-            && self.path.eq(&other.path)
-            && self.extensions.eq(&other.extensions)
-    }
-}
-
-impl ServerError {
-    /// Create a new server error with the message.
-    pub fn new(message: impl Into<String>, pos: Option<Pos>) -> Self {
-        Self {
-            message: message.into(),
-            source: None,
-            locations: pos.map(|pos| vec![pos]).unwrap_or_default(),
-            path: Vec::new(),
-            extensions: None,
-        }
-    }
-
-    /// Create a new server error with the message.
-    pub fn new_with_locations(message: impl Into<String>, locations: Vec<Pos>) -> Self {
-        Self {
-            message: message.into(),
-            source: None,
-            locations,
-            path: Vec::new(),
-            extensions: None,
-        }
-    }
-
-    /// Get the source of the error.
-    ///
-    /// # Examples
-    ///
-    /// ```rust, ignore
-    /// use engine::*;
-    /// use std::io::ErrorKind;
-    ///
-    /// struct Query;
-    ///
-    /// #[Object]
-    /// impl Query {
-    ///     async fn value(&self) -> Result<i32> {
-    ///         Err(Error::new_with_source(std::io::Error::new(ErrorKind::Other, "my error")))
-    ///     }
-    /// }
-    ///
-    /// let schema = Schema::new(Query, EmptyMutation, EmptySubscription);
-    ///
-    /// # tokio::runtime::Runtime::new().unwrap().block_on(async move {
-    /// let err = schema.execute("{ value }").await.into_result().unwrap_err().remove(0);
-    /// assert!(err.source::<std::io::Error>().is_some());
-    /// # });
-    /// ```
-    pub fn source<T: Any + Send + Sync>(&self) -> Option<&T> {
-        self.source.as_ref().and_then(|err| err.downcast_ref())
-    }
-
-    #[doc(hidden)]
-    #[must_use]
-    pub fn with_path(self, path: Vec<QueryPathSegment>) -> Self {
-        Self { path, ..self }
-    }
-}
-
-impl Display for ServerError {
-    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-        f.write_str(&self.message)
-    }
-}
-
-impl From<ServerError> for Vec<ServerError> {
-    fn from(single: ServerError) -> Self {
-        vec![single]
-    }
-}
-
-impl From<parser::Error> for ServerError {
-    fn from(e: parser::Error) -> Self {
-        Self {
-            message: e.to_string(),
-            source: None,
-            locations: e.positions().collect(),
-            path: Vec::new(),
-            extensions: None,
-        }
-    }
 }
 
 /// Alias for `Result<T, ServerError>`.
@@ -199,7 +59,7 @@ impl<T> InputValueError<T> {
 impl<T: LegacyInputType> InputValueError<T> {
     /// The expected input type did not match the actual input type.
     #[must_use]
-    pub fn expected_type(actual: Value) -> Self {
+    pub fn expected_type(actual: ConstValue) -> Self {
         Self::new(format!(
             r#"Expected input type "{}", found {}."#,
             T::type_name(),
@@ -245,7 +105,7 @@ impl<T: LegacyInputType, E: Display> From<E> for InputValueError<T> {
 pub type InputValueResult<T> = Result<T, InputValueError<T>>;
 
 /// An error with a message and optional extensions.
-#[derive(Clone, Serialize)]
+#[derive(Clone, serde::Serialize)]
 pub struct Error {
     /// The error message.
     pub message: String,
@@ -255,6 +115,10 @@ pub struct Error {
     /// Extensions to the error.
     #[serde(skip_serializing_if = "error_extensions_is_empty")]
     pub extensions: Option<ErrorExtensionValues>,
+}
+
+fn error_extensions_is_empty(values: &Option<ErrorExtensionValues>) -> bool {
+    values.as_ref().map_or(true, |values| values.0.is_empty())
 }
 
 impl Debug for Error {
